THRIFT-1986 fix TCompactProtocol constants

+++ lib/cpp/src/thrift/protocol/TCompactProtocol.h
@@ -35,10 +35,10 @@ class TCompactProtocolT
   : public TVirtualProtocol< TCompactProtocolT<Transport_> > {
 
  protected:
-  static const int8_t  PROTOCOL_ID = 0x82u;
+  static const int8_t  PROTOCOL_ID = (int8_t)0x82u;
   static const int8_t  VERSION_N = 1;
   static const int8_t  VERSION_MASK = 0x1f; // 0001 1111
-  static const int8_t  TYPE_MASK = 0xE0u; // 1110 0000
+  static const int8_t  TYPE_MASK = (int8_t)0xE0u; // 1110 0000
   static const int32_t TYPE_SHIFT_AMOUNT = 5;
 
   Transport_* trans_;
